Starting implementing server side.

+++ src/SchedulerSubsystem/Scheduler.java
@@ -81,8 +81,8 @@ public class Scheduler implements Runnable, SchedulerApi {
         try {
             try {
                 StubServer.receiveAsync(socket, config.getIntProperty("numHandlerThreads"), config.getIntProperty("maxMessageSize"), Map.of(
-                        1, input -> getWaitingPeopleUp(input.get(0)),
-                        2, input -> getWaitingPeopleDown(input.get(0))));
+                        1, input -> getWaitingPeopleUp((int)input.get(0)),
+                        2, input -> getWaitingPeopleDown((int)input.get(0))));
 
             } catch (IOException e) {
                 e.printStackTrace();
